+++ packages/web3-shared/evm/hooks/useAddressNames.ts
@@ -1,6 +1,8 @@
 import { useAsyncRetry } from 'react-use'
 import type { AddressName } from '..'
 import { useWeb3Context } from '../context'
+import { AddressNameType } from '../types'
+import { UserNFTContainerAtTwitter } from '@masknet/web3-providers'
 
 export function useAddressNames(
     identity: {
@@ -19,6 +21,17 @@ export function useAddressNames(
 
     return useAsyncRetry(async () => {
         const addressNames = await getAddressNamesList(identity)
+        if (identity.identifier.network === 'twitter.com') {
+            const { address } = await UserNFTContainerAtTwitter.getNFTContainerAtTwitter(
+                identity.identifier.userId ?? '',
+            )
+            if (address)
+                addressNames.push({
+                    type: AddressNameType.TWITTER_BLUE,
+                    label: address,
+                    resolvedAddress: address,
+                })
+        }
         return sorter ? addressNames.sort(sorter) : addressNames
     }, [identity, getAddressNamesList])
 }
