+++ packages/web3-shared/evm/package.json
@@ -10,6 +10,7 @@
     "@masknet/web3-constants": "workspace:*",
     "@masknet/web3-contracts": "workspace:*",
     "@masknet/web3-shared-base": "workspace:*",
+    "@masknet/web3-providers": "workspace:*",
     "@types/ethereum-blockies": "^0.1.0",
     "@types/punycode": "^2.1.0",
     "bignumber.js": "9.0.2",
